+++ CHANGELOG.md
@@ -12,6 +12,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add method `MonitoredItemBuilder::attribute_id()` to set a different attribute ID to monitor.
 - Add method `MonitoredItemBuilder::filter()` and related data types `ua::DataChangeFilter`,
   `ua::EventFilter`, `ua::AggregateFilter` along with associated data types.
+- Add methods `ClientBuilder::security_mode()` and `ClientBuilder::security_policy_uri()` to select
+  security mode and security policy URI when filtering remote endpoints.
 
 ### Changed